2022.07.31 added abstract classes, overrided tostring, equals, hashcode hw3

+++ src/root/human/doctor/Dentist.java
@@ -1,21 +1,15 @@
 package root.human.doctor;
-
 import root.human.properties.Address;
 import root.human.properties.Phone;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Dentist extends Doctor {
-
     private String tooth;
     public static int countDentist;
 
-
-
-
-
-
-   public Dentist(LocalDate dateOfBirth, String name, Address address, Phone phone, String specialty, int price) {
+   public Dentist(LocalDate dateOfBirth, String name, Address address, Phone phone, String specialty, BigDecimal price) {
         super(dateOfBirth, name, address, phone, specialty, price);
         countDentist++;
     }
