+++ src/root/human/patient/prescription/procedure/Procedure.java
@@ -1,13 +1,15 @@
 package root.human.patient.prescription.procedure;
 
+import java.math.BigDecimal;
+
 public class Procedure {
-    //Fields
+
     private String name;
     private String description;
-    private int price;
+    private BigDecimal price;
     private int duration;
 
-    public Procedure(String name, String description, int price, int quantity) {
+    public Procedure(String name, String description, BigDecimal price, int quantity) {
         this.name = name;
         this.description = description;
         this.price = price;
@@ -35,11 +37,11 @@ public class Procedure {
         this.description = description;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
