+++ src/root/human/patient/prescription/procedure/ProcedureExternal.java
@@ -1,7 +1,9 @@
 package root.human.patient.prescription.procedure;
 
+import java.math.BigDecimal;
+
 public class ProcedureExternal extends Procedure{
-    public ProcedureExternal(String name, String description, int price, int quantity) {
+    public ProcedureExternal(String name, String description, BigDecimal price, int quantity) {
         super(name, description, price, quantity);
     }
 }
